requirements.txt
flask
openai
python-dotenv
gunicorn
pydub